+++ modules/checkout/views/payment/pagofacil-payment-view.php
@@ -1,5 +1,6 @@
 <?php
 
+use app\modules\checkout\models\PagoFacilTransmitionFile;
 use yii\bootstrap\Html;
 use yii\grid\GridView;
 use yii\helpers\Url;
@@ -47,6 +48,37 @@ $this->params['breadcrumbs'][] = 'Pago Fácil - ' . $model->upload_date;
         ],       
       ]);
     ?>
+
+    <!-- Inicio de Progress Bar -->
+    <?php if ($model->status == PagoFacilTransmitionFile::STATUS_PENDING) { ?>
+        <div class="panel panel-default" id="panel-progress">
+            <div class="panel-heading">
+                <h3 class="panel-title"><?= Yii::t('app', 'Progress') ?></h3>
+            </div>
+            <div class="panel-body collapse in" id="panel-body-progress" aria-expanded="true">
+                <div class="row">
+                    <div class="col-sm-5">
+                        <div class="row">
+                            <div class="col-sm-6"><?php echo Yii::t('app', 'Total to Process') ?></div>
+                            <div class="col-sm-6" id="total_to_process"></div>
+                        </div>
+                        <div class="row">
+                            <div class="col-sm-6"><?php echo Yii::t('app', 'Processed') ?></div>
+                            <div class="col-sm-6" id="processed"></div>
+                        </div>
+                    </div>
+                    <div class="col-sm-7">
+                        <div class="progress">
+                            <div id="progress-bar" class="progress-bar" role="progressbar" aria-valuenow="0"
+                                 aria-valuemin="0" aria-valuemax="100">
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    <?php } ?>
+    <!-- Fin de Progress Bar -->
 
     <?php
         $columns[] = ['class' => 'yii\grid\SerialColumn'];
@@ -119,13 +151,69 @@ $this->params['breadcrumbs'][] = 'Pago Fácil - ' . $model->upload_date;
 <script>
 
     var PagoFacilView= new function(){
+        this.processing = false;
+
         this.init= function(){
            $(document).on('click', '#confirm', function(e){
                 e.preventDefault();
                 $('#confirm').html('Procesando...');
                 PagoFacilView.confirm();
-           }); 
+           });
+
+           PagoFacilView.checkIfProcessItsStarted();
         
+        }
+
+        this.checkIfProcessItsStarted = function () {
+            $.ajax({
+                url: '<?= Url::to(["close-payment-process-started"])?>',
+                method: 'GET',
+                data: {pago_facil_transmition_file_id : <?= $model->pago_facil_transmition_file_id?> },
+                datatType: 'json',
+                success: function (data) {
+                    console.log(data);
+                    if(data.process_started) {
+                        console.log('started');
+                        $('#panel-progress').show();
+                        $('#panel-filtro').hide();
+                        PagoFacilView.processing = true;
+                        setTimeout(PagoFacilView.getProceso(), 500);
+                    } else {
+                        console.log('no started');
+                        PagoFacilView.processing = false;
+                    }
+                }
+            })
+        }
+
+        this.getProceso = function() {
+            setTimeout(function(){
+                $.ajax({
+                    method: 'POST',
+                    url: '<?=Url::to(['get-close-payment-process'])?>',
+                    dataType: 'json',
+                    success: function(data, textStatus, jqXhr) {
+                        console.log(data);
+                        var value = ((data.qty*100)/data.total);
+                        $('.progress-bar').css('width', value+'%').attr('aria-valuenow', value);
+                        $('#total_to_process').html(data.total);
+                        $('#processed').html(data.qty);
+                        if(data.total!=data.qty) {
+                            $('.progress-bar').html(parseInt( value) +'%');
+                        } else {
+                            $('.progress-bar').html('<?php echo Yii::t('app', 'Process finished') ?>');
+                            $('#process-label').addClass('hidden');
+                            if(data.total != 0 && data.qty != 0) {
+                                PagoFacilView.processing = false;
+                            }
+                        }
+
+                        if( PagoFacilView.processing ) {
+                            PagoFacilView.getProceso();
+                        }
+                    }
+                });
+            }, 2000)
         }
         
         this.confirm= function(){